basic simulation is now working and producing a fitness value, truck implementation is still bogus

+++ evo/src/main/java/rinde/evo4mas/evo/gp/GPComputationResult.java
@@ -11,6 +11,6 @@ import rinde.cloud.javainterface.ComputationResult;
  */
 public interface GPComputationResult extends ComputationResult {
 
-	double getFitness();
+	float getFitness();
 
 }
